requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
chromadb
faiss-cpu
SpeechRecognition
pydub
plotly
seaborn
